refactor: reduce repeated code of pickers
