+++ packages/vrender-core/src/picker/pick-interceptor.ts
@@ -1,9 +1,7 @@
 import type { IMatrix, IPointLike } from '@visactor/vutils';
-import { Point, pi2 } from '@visactor/vutils';
+import { Point } from '@visactor/vutils';
 import { injectable } from '../common/inversify-lite';
-import { ARC3D_NUMBER_TYPE } from '../graphic/constants';
 import type {
-  IArc,
   IContext2d,
   IGraphic,
   IGroup,
@@ -13,6 +11,7 @@ import type {
   PickResult
 } from '../interface';
 import { matrixAllocate } from '../allocator/matrix-allocate';
+import { draw3dItem } from '../common/3d-interceptor';
 
 // 拦截器
 export const PickItemInterceptor = Symbol.for('PickItemInterceptor');
@@ -164,111 +163,14 @@ export class Canvas3DPickItemInterceptor implements IPickItemInterceptorContribu
 
     // 设置context的transform到上一个节点
     if (graphic.isContainer) {
-      // hack逻辑，如果是饼图的话，需要依次绘制不同的边
-      let isPie: boolean = false;
-      let is3d: boolean = false;
-      graphic.forEachChildren((c: IGraphic) => {
-        isPie = c.numberType === ARC3D_NUMBER_TYPE;
-        return !isPie;
-      });
-      graphic.forEachChildren((c: IGraphic) => {
-        is3d = !!c.findFace;
-        return !is3d;
-      });
-
-      let result: PickResult;
-      if (isPie) {
-        const children = graphic.getChildren() as IArc[];
-        // 绘制内层
-        // drawContext.hack_pieFace = 'inside';
-        // drawContribution.renderGroup(graphic as IGroup, drawContext);
-        // 绘制底部
-        // drawContext.hack_pieFace = 'bottom';
-        // drawContribution.renderGroup(graphic as IGroup, drawContext);
-        // 绘制外部
-        // 排序一下
-        const sortedChildren = [...children];
-        sortedChildren.sort((a, b) => {
-          let angle1 = (a.attribute.startAngle ?? 0 + a.attribute.endAngle ?? 0) / 2;
-          let angle2 = (b.attribute.startAngle ?? 0 + b.attribute.endAngle ?? 0) / 2;
-          while (angle1 < 0) {
-            angle1 += pi2;
-          }
-          while (angle2 < 0) {
-            angle2 += pi2;
-          }
-          return angle2 - angle1;
-        });
-        sortedChildren.forEach(c => {
-          c._next = null;
-          c._prev = null;
-        });
-        graphic.removeAllChild();
-        graphic.update();
-        sortedChildren.forEach(c => {
-          graphic.appendChild(c);
-        });
-        pickParams.hack_pieFace = 'outside';
-        result = pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
-        if (!result.graphic) {
-          // 绘制内部
-          pickParams.hack_pieFace = 'inside';
-          result = pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
-        }
-        if (!result.graphic) {
-          // 绘制顶部
-          pickParams.hack_pieFace = 'top';
-          result = pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
-        }
-        graphic.removeAllChild();
-        children.forEach(c => {
-          c._next = null;
-          c._prev = null;
-        });
-        children.forEach(c => {
-          graphic.appendChild(c);
-        });
-      } else if (is3d) {
-        // 排序这些图元
-        const children = graphic.getChildren() as IGraphic[];
-        const zChildren = children.map(g => {
-          const face3d = g.findFace();
-          const vertices = face3d.vertices;
-          // 计算每个顶点的view
-          const viewdVerticesZ = vertices.map(v => {
-            return context.view(v[0], v[1], v[2] + g.attribute.z ?? 0)[2];
-          });
-          const ave_z = viewdVerticesZ.reduce((a, b) => a + b, 0);
-          return {
-            ave_z,
-            g
-          };
-        });
-        zChildren.sort((a, b) => b.ave_z - a.ave_z);
-        graphic.removeAllChild();
-        zChildren.forEach(i => {
-          i.g._next = null;
-          i.g._prev = null;
-        });
-        graphic.update();
-        zChildren.forEach(i => {
-          graphic.add(i.g);
-        });
-
-        result = pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
-
-        graphic.removeAllChild();
-        children.forEach(g => {
-          g._next = null;
-          g._prev = null;
-        });
-        graphic.update();
-        children.forEach(g => {
-          graphic.add(g);
-        });
-      } else {
-        result = pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
-      }
+      const result = draw3dItem(
+        context,
+        graphic,
+        () => {
+          return pickerService.pickGroup(graphic as IGroup, point, params.parentMatrix, pickParams);
+        },
+        pickParams
+      );
 
       context.camera = null;
