+++ packages/vrender-core/src/render/contributions/render/arc-render.ts
@@ -97,9 +97,7 @@ export class DefaultCanvasArcRender extends BaseRender<IArc> implements IGraphic
     let collapsedToLine: boolean = false;
     // 规范化outerRadius和innerRadius
     if (outerRadius < innerRadius) {
-      const temp = outerRadius;
-      outerRadius = innerRadius;
-      innerRadius = temp;
+      [innerRadius, outerRadius] = [outerRadius, innerRadius];
     }
 
     const cornerRadius = arc.getParsedCornerRadius();
