+++ packages/vrender-core/src/render/contributions/render/draw-interceptor.ts
@@ -1,9 +1,8 @@
 import { injectable } from '../../../common/inversify-lite';
-import { AABBBounds, pi2 } from '@visactor/vutils';
+import { AABBBounds } from '@visactor/vutils';
 import { mat3Tomat4, multiplyMat4Mat4 } from '../../../graphic/graphic-service/graphic-service';
 import { graphicCreator } from '../../../graphic/graphic-creator';
 import type {
-  IArc,
   IContext2d,
   IDrawContext,
   IDrawContribution,
@@ -15,7 +14,7 @@ import type {
   IRenderService
 } from '../../../interface';
 import { mat4Allocate } from '../../../allocator/matrix-allocate';
-import { ARC3D_NUMBER_TYPE } from '../../../graphic/constants';
+import { draw3dItem } from '../../../common/3d-interceptor';
 
 // 拦截器
 export const DrawItemInterceptor = Symbol.for('DrawItemInterceptor');
@@ -427,106 +426,19 @@ export class Canvas3DDrawItemInterceptor implements IDrawItemInterceptorContribu
 
     // 设置context的transform到上一个节点
     if (graphic.isContainer) {
-      // hack逻辑，如果是饼图的话，需要依次绘制不同的边
-      let isPie: boolean = false;
-      let is3d: boolean = false;
-      graphic.forEachChildren((c: IGraphic) => {
-        isPie = c.numberType === ARC3D_NUMBER_TYPE;
-        return !isPie;
-      });
-      graphic.forEachChildren((c: IGraphic) => {
-        is3d = !!c.findFace;
-        return !is3d;
-      });
-      if (isPie) {
-        const children = graphic.getChildren() as IArc[];
-        // 绘制内层
-        // drawContext.hack_pieFace = 'inside';
-        // drawContribution.renderGroup(graphic as IGroup, drawContext);
-        // 绘制底部
-        // drawContext.hack_pieFace = 'bottom';
-        // drawContribution.renderGroup(graphic as IGroup, drawContext);
-        // 绘制外部
-        // 排序一下
-        const sortedChildren = [...children];
-        sortedChildren.sort((a, b) => {
-          let angle1 = ((a.attribute.startAngle ?? 0) + (a.attribute.endAngle ?? 0)) / 2;
-          let angle2 = ((b.attribute.startAngle ?? 0) + (b.attribute.endAngle ?? 0)) / 2;
-          while (angle1 < 0) {
-            angle1 += pi2;
-          }
-          while (angle2 < 0) {
-            angle2 += pi2;
-          }
-          return angle2 - angle1;
-        });
-        sortedChildren.forEach(c => {
-          c._next = null;
-          c._prev = null;
-        });
-        graphic.removeAllChild();
-        graphic.update();
-        sortedChildren.forEach(c => {
-          graphic.appendChild(c);
-        });
-        const m = graphic.parent.globalTransMatrix;
-        drawContext.hack_pieFace = 'outside';
-        drawContribution.renderGroup(graphic as IGroup, drawContext, m);
-        // 绘制内部
-        drawContext.hack_pieFace = 'inside';
-        drawContribution.renderGroup(graphic as IGroup, drawContext, m);
-        // 绘制顶部
-        drawContext.hack_pieFace = 'top';
-        drawContribution.renderGroup(graphic as IGroup, drawContext, m);
-        graphic.removeAllChild();
-        children.forEach(c => {
-          c._next = null;
-          c._prev = null;
-        });
-        children.forEach(c => {
-          graphic.appendChild(c);
-        });
-      } else if (is3d) {
-        // 排序这些图元
-        const children = graphic.getChildren() as IGraphic[];
-        const zChildren = children.map(g => {
-          const face3d = g.findFace();
-          const vertices = face3d.vertices;
-          // 计算每个顶点的view
-          const viewdVerticesZ = vertices.map(v => {
-            return context.view(v[0], v[1], v[2] + g.attribute.z ?? 0)[2];
-          });
-          const ave_z = viewdVerticesZ.reduce((a, b) => a + b, 0);
-          return {
-            ave_z,
-            g
-          };
-        });
-        zChildren.sort((a, b) => b.ave_z - a.ave_z);
-        graphic.removeAllChild();
-        zChildren.forEach(i => {
-          i.g._next = null;
-          i.g._prev = null;
-        });
-        graphic.update();
-        zChildren.forEach(i => {
-          graphic.add(i.g);
-        });
-
-        drawContribution.renderGroup(graphic as IGroup, drawContext, graphic.parent.globalTransMatrix, true);
-
-        graphic.removeAllChild();
-        children.forEach(g => {
-          g._next = null;
-          g._prev = null;
-        });
-        graphic.update();
-        children.forEach(g => {
-          graphic.add(g);
-        });
-      } else {
-        drawContribution.renderGroup(graphic as IGroup, drawContext, graphic.parent.globalTransMatrix);
-      }
+      draw3dItem(
+        context,
+        graphic,
+        (isPie: boolean, is3d: boolean) => {
+          return drawContribution.renderGroup(
+            graphic as IGroup,
+            drawContext,
+            graphic.parent.globalTransMatrix,
+            !isPie && is3d
+          );
+        },
+        drawContext
+      );
     } else {
       drawContribution.renderItem(graphic, drawContext);
     }
